+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.2] - 2025-05-29
+
+### Fixed
+
+- Support for extra animal config produce overrides + extra drops
+- Fixed compatiblity problem with Animal Squeeze Through + EAC extra houses
+
 ## [1.2.1] - 2025-04-13
 
 ### Added